Feyz

+++ cms-project-web/src/globalClasses/WorkoutDetails.js
@@ -1,11 +1,21 @@
 import React from "react";
 import { useWorkoutsContext } from "../Hook/useWorkoutsContext";
+import formatDistanceToNow from "date-fns/formatDistanceToNow";
+import { useAuthContext } from "../Hook/useAuthContext";
 
 const WorkoutDetails = ({ workout }) => {
+  const { user } = useAuthContext();
   const { dispatch } = useWorkoutsContext();
   const handleClick = async () => {
+    if (!user) {
+      return;
+    }
+
     const response = await fetch("api/workouts/" + workout._id, {
       method: "DELETE",
+      headers: {
+        Authorization: `Bearer ${user.token}`,
+      },
     });
 
     const json = await response.json();
@@ -18,15 +28,19 @@ const WorkoutDetails = ({ workout }) => {
     <div className="workout-details">
       <h4>{workout.title}</h4>
       <p>
-        <strong>Load (kg): </strong>
+        <strong>Id: </strong>
         {workout.load}
       </p>
       <p>
-        <strong>Number of reps: </strong>
+        <strong>Condact Number: </strong>
         {workout.reps}
       </p>
-      <p>{workout.createdAt}</p>
-      <span onClick={handleClick}>Delete</span>
+      <p>
+        {formatDistanceToNow(new Date(workout.createdAt), { addSuffix: true })}
+      </p>
+      <span className="material-symbols-outlined" onClick={handleClick}>
+        Delete
+      </span>
     </div>
   );
 };
